Updated babel to use interface to props plugin
