+++ package.json
@@ -64,6 +64,7 @@
     "bootstrap": "lerna bootstrap",
     "build": "yarn clean:build && lerna run prebuild --parallel && lerna run build --parallel && lerna run postbuild",
     "build:docs": "yarn build && lerna run docbuild --scope=@patternfly/react-docs --stream",
+    "build:typescript": "lerna run build --scope=@patternfly/react-core --stream",
     "build:prdocs": "lerna run pr-build --scope=@patternfly/react-docs --stream",
     "clean": "yarn clean:build",
     "clean:build": "rm -rf packages/*/dist && rm -rf packages/patternfly-?/*/dist",
