+++ packages/patternfly-4/react-core/tsconfig.json
@@ -1,9 +1,8 @@
 {
   "compilerOptions": {
+    "moduleResolution": "node",
     "skipLibCheck": true,
-    "outDir": "./tsc_out/", // path to output directory
     "sourceMap": true, // allow sourcemap support
-    "strictNullChecks": true, // enable strict null checks as a best practice
     "lib": ["es6", "dom"],
     "module": "es2015", // specify module code generation
     "jsx": "react", // use typescript to transpile jsx to js
@@ -11,5 +10,5 @@
     "allowJs": true // allow a partial TypeScript and JavaScript codebase
   },
   "include": ["./src/**/*"],
-  "exclude": ["./node_modules", "./src/**/*.d.ts"]
+  "exclude": ["./src/**/*.d.ts"]
 }
